Listen to the parameter changes in the ActorFilter
When setting the array of filters on the layer, the filters are copied which means that parameter changes only get set on the layer's copy and not the original.
Therefore when setting a new array of filters (on filter addition or removal) the filters will use the original values rather than the modified ones.
To solve this we listen to the param changes in the ActorFilter as well, and set them on the original copy of the CIFilter