Create a layout manager for the bottom layout
Layout manager that will arrange the children boxes horizontally in order. Fixed width views will not be changed in size, the rest of the space will be divided evenly between the rest of the children.
Assigns itself as a delegate for the box, and animates frame changes for the other items when a box is opened or closed.